requirements.txt
Flask
flask-cors
Pillow
openai
python-dotenv
python-multipart
